EK-11: Refactored map style helper, divided code into smaller helpers
